refactor(lints): Move LINT_GROUPS next to LINTS
