+++ src/cargo/lints/mod.rs
@@ -18,7 +18,7 @@ pub mod rules;
 
 pub use lint::{Lint, LintGroup, LintLevel, LintLevelSource};
 pub use report::{AsIndex, get_key_value, get_key_value_span, rel_cwd_manifest_path};
-pub use rules::LINTS;
+pub use rules::{LINT_GROUPS, LINTS};
 
 /// Scope at which a lint runs: package-level or workspace-level.
 pub enum ManifestFor<'a> {
@@ -196,217 +196,4 @@ fn report_feature_not_enabled(
     gctx.shell().print_report(&report, true)?;
 
     Ok(())
-}
-
-pub static LINT_GROUPS: &[LintGroup] = &[
-    COMPLEXITY,
-    CORRECTNESS,
-    NURSERY,
-    PEDANTIC,
-    PERF,
-    RESTRICTION,
-    STYLE,
-    SUSPICIOUS,
-    TEST_DUMMY_UNSTABLE,
-];
-
-const COMPLEXITY: LintGroup = LintGroup {
-    name: "complexity",
-    desc: "code that does something simple but in a complex way",
-    default_level: LintLevel::Warn,
-    feature_gate: None,
-    hidden: false,
-};
-
-const CORRECTNESS: LintGroup = LintGroup {
-    name: "correctness",
-    desc: "code that is outright wrong or useless",
-    default_level: LintLevel::Deny,
-    feature_gate: None,
-    hidden: false,
-};
-
-const NURSERY: LintGroup = LintGroup {
-    name: "nursery",
-    desc: "new lints that are still under development",
-    default_level: LintLevel::Allow,
-    feature_gate: None,
-    hidden: false,
-};
-
-const PEDANTIC: LintGroup = LintGroup {
-    name: "pedantic",
-    desc: "lints which are rather strict or have occasional false positives",
-    default_level: LintLevel::Allow,
-    feature_gate: None,
-    hidden: false,
-};
-
-const PERF: LintGroup = LintGroup {
-    name: "perf",
-    desc: "code that can be written to run faster",
-    default_level: LintLevel::Warn,
-    feature_gate: None,
-    hidden: false,
-};
-
-const RESTRICTION: LintGroup = LintGroup {
-    name: "restriction",
-    desc: "lints which prevent the use of Cargo features",
-    default_level: LintLevel::Allow,
-    feature_gate: None,
-    hidden: false,
-};
-
-const STYLE: LintGroup = LintGroup {
-    name: "style",
-    desc: "code that should be written in a more idiomatic way",
-    default_level: LintLevel::Warn,
-    feature_gate: None,
-    hidden: false,
-};
-
-const SUSPICIOUS: LintGroup = LintGroup {
-    name: "suspicious",
-    desc: "code that is most likely wrong or useless",
-    default_level: LintLevel::Warn,
-    feature_gate: None,
-    hidden: false,
-};
-
-/// This lint group is only to be used for testing purposes
-const TEST_DUMMY_UNSTABLE: LintGroup = LintGroup {
-    name: "test_dummy_unstable",
-    desc: "test_dummy_unstable is meant to only be used in tests",
-    default_level: LintLevel::Allow,
-    feature_gate: Some(Feature::test_dummy_unstable()),
-    hidden: true,
-};
-
-#[cfg(test)]
-mod tests {
-    use itertools::Itertools;
-    use snapbox::ToDebug;
-    use std::collections::HashSet;
-
-    #[test]
-    fn ensure_lint_groups_do_not_default_to_forbid() {
-        let forbid_groups = super::LINT_GROUPS
-            .iter()
-            .filter(|g| matches!(g.default_level, super::LintLevel::Forbid))
-            .collect::<Vec<_>>();
-
-        assert!(
-            forbid_groups.is_empty(),
-            "\n`LintGroup`s should never default to `forbid`, but the following do:\n\
-            {}\n",
-            forbid_groups.iter().map(|g| g.name).join("\n")
-        );
-    }
-
-    #[test]
-    fn ensure_sorted_lints() {
-        // This will be printed out if the fields are not sorted.
-        let location = std::panic::Location::caller();
-        println!("\nTo fix this test, sort `LINTS` in {}\n", location.file(),);
-
-        let actual = super::LINTS
-            .iter()
-            .map(|l| l.name.to_uppercase())
-            .collect::<Vec<_>>();
-
-        let mut expected = actual.clone();
-        expected.sort();
-        snapbox::assert_data_eq!(actual.to_debug(), expected.to_debug());
-    }
-
-    #[test]
-    fn ensure_sorted_lint_groups() {
-        // This will be printed out if the fields are not sorted.
-        let location = std::panic::Location::caller();
-        println!(
-            "\nTo fix this test, sort `LINT_GROUPS` in {}\n",
-            location.file(),
-        );
-        let actual = super::LINT_GROUPS
-            .iter()
-            .map(|l| l.name.to_uppercase())
-            .collect::<Vec<_>>();
-
-        let mut expected = actual.clone();
-        expected.sort();
-        snapbox::assert_data_eq!(actual.to_debug(), expected.to_debug());
-    }
-
-    #[test]
-    fn ensure_updated_lints() {
-        let dir = snapbox::utils::current_dir!().join("rules");
-        let mut expected = HashSet::new();
-        for entry in std::fs::read_dir(&dir).unwrap() {
-            let entry = entry.unwrap();
-            let path = entry.path();
-            if path.ends_with("mod.rs") {
-                continue;
-            }
-            let lint_name = path.file_stem().unwrap().to_string_lossy();
-            assert!(expected.insert(lint_name.into()), "duplicate lint found");
-        }
-
-        let actual = super::LINTS
-            .iter()
-            .map(|l| l.name.to_string())
-            .collect::<HashSet<_>>();
-        let diff = expected.difference(&actual).sorted().collect::<Vec<_>>();
-
-        let mut need_added = String::new();
-        for name in &diff {
-            need_added.push_str(&format!("{name}\n"));
-        }
-        assert!(
-            diff.is_empty(),
-            "\n`LINTS` did not contain all `Lint`s found in {}\n\
-            Please add the following to `LINTS`:\n\
-            {need_added}",
-            dir.display(),
-        );
-    }
-
-    #[test]
-    fn ensure_updated_lint_groups() {
-        let path = snapbox::utils::current_rs!();
-        let expected = std::fs::read_to_string(&path).unwrap();
-        let expected = expected
-            .lines()
-            .filter_map(|l| {
-                if l.ends_with(": LintGroup = LintGroup {") {
-                    Some(
-                        l.chars()
-                            .skip(6)
-                            .take_while(|c| *c != ':')
-                            .collect::<String>(),
-                    )
-                } else {
-                    None
-                }
-            })
-            .collect::<HashSet<_>>();
-        let actual = super::LINT_GROUPS
-            .iter()
-            .map(|l| l.name.to_uppercase())
-            .collect::<HashSet<_>>();
-        let diff = expected.difference(&actual).sorted().collect::<Vec<_>>();
-
-        let mut need_added = String::new();
-        for name in &diff {
-            need_added.push_str(&format!("{}\n", name));
-        }
-        assert!(
-            diff.is_empty(),
-            "\n`LINT_GROUPS` did not contain all `LintGroup`s found in {}\n\
-            Please add the following to `LINT_GROUPS`:\n\
-            {}",
-            path.display(),
-            need_added
-        );
-    }
 }
