+++ src/cargo/lints/rules/im_a_teapot.rs
@@ -8,13 +8,13 @@ use cargo_util_terminal::report::Origin;
 use cargo_util_terminal::report::Snippet;
 use tracing::instrument;
 
+use super::TEST_DUMMY_UNSTABLE;
 use crate::CargoResult;
 use crate::GlobalContext;
 use crate::core::Feature;
 use crate::core::Package;
 use crate::lints::Lint;
 use crate::lints::LintLevel;
-use crate::lints::TEST_DUMMY_UNSTABLE;
 use crate::lints::get_key_value_span;
 use crate::lints::rel_cwd_manifest_path;
