+++ src/cargo/lints/rules/non_kebab_case_bins.rs
@@ -9,6 +9,7 @@ use cargo_util_terminal::report::Patch;
 use cargo_util_terminal::report::Snippet;
 use tracing::instrument;
 
+use super::STYLE;
 use crate::CargoResult;
 use crate::GlobalContext;
 use crate::core::Package;
@@ -17,7 +18,6 @@ use crate::lints::AsIndex;
 use crate::lints::Lint;
 use crate::lints::LintLevel;
 use crate::lints::LintLevelSource;
-use crate::lints::STYLE;
 use crate::lints::get_key_value_span;
 use crate::lints::rel_cwd_manifest_path;
