+++ src/cargo/lints/rules/unused_dependencies.rs
@@ -10,12 +10,12 @@ use cargo_util_terminal::report::Patch;
 use cargo_util_terminal::report::Snippet;
 use tracing::instrument;
 
+use super::STYLE;
 use crate::CargoResult;
 use crate::GlobalContext;
 use crate::core::Package;
 use crate::lints::Lint;
 use crate::lints::LintLevel;
-use crate::lints::STYLE;
 use crate::lints::get_key_value_span;
 use crate::lints::rel_cwd_manifest_path;
