+++ src/cargo/lints/rules/unused_workspace_dependencies.rs
@@ -11,13 +11,13 @@ use cargo_util_terminal::report::Snippet;
 use indexmap::IndexSet;
 use tracing::instrument;
 
+use super::SUSPICIOUS;
 use crate::CargoResult;
 use crate::GlobalContext;
 use crate::core::MaybePackage;
 use crate::core::Workspace;
 use crate::lints::Lint;
 use crate::lints::LintLevel;
-use crate::lints::SUSPICIOUS;
 use crate::lints::get_key_value_span;
 use crate::lints::rel_cwd_manifest_path;
